// src/shared/code-erros/error-codes.enum.ts
export enum ErrorCodesEnum {
  MUS000 = 'MUS000',
  MUS001 = 'MUS001',
  MUS002 = 'MUS002',
  MUS003 = 'MUS003',
  MUS004 = 'MUS004',
  MUS005 = 'MUS005',
  MUS006 = 'MUS006',
  MUS007 = 'MUS007',
  MUS008 = 'MUS008',
  MUS009 = 'MUS009',
  MUS010 = 'MUS010',
  MUS011 = 'MUS011',
  MUS012 = 'MUS012',
  MUS013 = 'MUS013',
  MUS014 = 'MUS014',
  MUS015 = 'MUS015',
  MUS016 = 'MUS016',
  MUS017 = 'MUS017',
  MUS018 = 'MUS018',
  MUS019 = 'MUS019',
  MUS020 = 'MUS020',
  MUS021 = 'MUS021',
  MUS022 = 'MUS022',
  MUS023 = 'MUS023',
  MUS024 = 'MUS024',
  MUS025 = 'MUS025',
}